Improve i18n to avoid the so-called "Lego message construction"
This allows for more flexible i18n as translators can now control the placement of things like the dice roll result in the "youRolled" message.
Previously only the string "You rolled" was translatable; a space, followed by the string representing the number, was automatically appended to this.
The space might be incorrect/not needed for certain languages (e.g. CJK languages) and it's also totally possible that languages may want to order the words differently than English/Finnish/Swedish.

* The message representing the dice roll result is passed as $1 to the "youRolled" message
* Player name is passed as $1 to the "wonGame" message
* Messages "still" and "turnsLeft" were stitched together and split into two messages, "stillOneTurnLeft" and "stillTurnsLeft" (for singular and plural, respectively)
* Version number and build date bumped, 'cause why not